Move preview components to separate files
